#18 - Established Win&Lose conditions and events.
IKGameModeBase - As well as a hero container, populated a new container for enemies. Added a function that will remove an actor from each containers. Created events when win or lose. However, there are not enough implementation on call the functions when actors died. Need changes in BP_Hero, and BP_Enemy.