#21 - Successfully resolve the old issue, change idiom of max stats.
BP_Cover - In its CharacterStatComponent, Changed a character ID to 1, which means cover.
BP_Enemy, BP_Hero - Add HP widgets properly. In order to add the UI, need to bind its character stat to the widget in BeginPlay().
BPHitPointsUI - A new version of HitPointUI.
IK_Proto_GameData - Add a new field that represents status of covers.
HitPoint 60.
CharacterStatComponent - Slightly changed the meaning of max attributes. They are no longer has meaning of theoretical limitations. it means initial data of each attributes.
HitPointsUI - It is a class that has logical implementation of old HP UI. Now they are in C++ code, but visual implementations are still on the BP.