#24 - Successfully adapt item systems.
BP_ButtonBarWidget - A blueprint file that derived by ButtonBarWidget class.
BP_IKHUD - Change HUD widget.
ButtonBarWidget, BP_SkillBarWidget, ItemBarWidget - Because of an issue that is binding multiple functions on a delegator, merge SkillBarWidget and ItemBarWidget. Renamed it ButtonBarWidget.
IKGameInstance - Add an item into inventory when other resources initialized (From ctor to BeginPlay)
MyTestItem - Print a log when the item is used.
TargetingComponent - Add a new function that is called to use item. It does not take a pointer to invoker.