#37 - Implement a UI to visualize maps.
BP_MyMapWidget - A BP derived by MapWidget.
IKGameInstance - Created a getter to the IKMaps
IKMaps - Added UENUM and USTRUCT on Node resources. Added interfaces to get map data.
MapWidget - A new class that is derived by UUserWidget. However its button is not visible. I believe the canvas panel is not yet added properly. A chunk of tasks is not implemented yet in here.