#37 - Successfully displayed paths in Map widget.
MapWidget - Draw lines if and only if any of buttons in link is valid. Add clipping on the line. The line is only drawn in the scroll box space. Use GetPaintSpaceGeometry instead of GetCachedGeometry function to improve stability. GetCachedGeometry caused visual errors in Stand-alone mode.

TODO: Set scroll offset based on the map state.