#55 - Removed FObjectLoader. Applied BP.
BP_IKGameInstance - Set BP_ItemInventory class
BP_Item, BP_ItemInventory, BP_BP_SmokeRegion - Created BP child classes of each of them.
Item - Added TSubclassOf<ASmokeRegion> to instantiate BP of it.
ItemInventory - Added TSubclassOf<UItem> to instantiate BP of it.
SmokeRegion - In order to use of BP Material, added UMaterialInstance with EditAnyuwhere, and BlueprintRreadWrite. Since assigning BP to pointer is done after Natice constructor of the class, it is nullptr in Constructor. Thus, should check nullptr and use of it in BeginPlay() or PostInitializeComponents().
IKGameInstance - Added TSubclassOf<UItemInventory> to do it.