#96 - Create rough visuals of ItemKeepOrDiscardUI.
BP_ItemInventory - Move a UI to UI/Items folder.
BP_CheckboxButtonWidget, BP_ItemKeepOrDiscardWidget - Blueprint assets.
BP_IKGotchaGameMode, BP_IKGotchaHUD - Temporarily display ItemKeepDiscardWidget to test purpose.
CheckboxButtonWidget - Since built-in check box is not enough for my purpose. Thus, I need to implement a custom checkbox button widget to display it is checked or not.
ItemKeepOrDiscardWidget - When a new item came to the item inventory, users need to choose which one to keep or discard. It is a UI to do it.
IKGotchaHUD - Display a blueprintly determined UI.
IKGotchaPlayerController - Show mouse cursor and set appropriate input mode.